+++ grass_project/scene_object_arrays.cpp
@@ -15,7 +15,7 @@ void SceneObjectArrays::createVertexArray(const std::vector<float>& positions) {
 	createArrayBuffer(positions);
 	GLCall(glEnableVertexAttribArray(0));
 	GLCall(glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 3 * sizeof(float), (void*)0));
-	vertexCount = positions.size();
+	vertexCount = (unsigned int)positions.size();
 }
 
 void SceneObjectArrays::draw(Scene& scene) {
